RDKCOM-3318: REFPLTB-1842 test-and-diagnostic build failure in RPi4 Broadban...
build errors in test-and-diagnostic while building RPi4 Broadband 64bit:
| TestAndDiagnostic/source/dmltad/cosa_wanconnectivity_operations.c:1227:24: error: format '%lld' expects argument of type 'long long int', but argument 4 has type 'int64_t' {aka 'long int'} [-Werror=format=]
|  1227 |         WANCHK_LOG_DBG(waiting for response with timeout:%lld\n,timeout);
|       |                                                               ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ~~~~~~~
|       |                                                               |
|       |                                                               int64_t {aka long int}
| TestAndDiagnostic/source/dmltad/cosa_wanconnectivity_apis.h:103:85: note: in definition of macro 'WANCHK_LOG_DBG'
|   103 | #define WANCHK_LOG_DBG(fmt...)    RDK_LOG(RDK_LOG_DEBUG,   LOG.RDK.WANCNCTVTYCHK, fmt)
|       |                                                                                     ^~~
| TestAndDiagnostic/source/dmltad/cosa_wanconnectivity_operations.c:1227:62: note: format string is defined here
|  1227 |         WANCHK_LOG_DBG(waiting for response with timeout:%lld\n,timeout);
|       |                                                           ~~~^
|       |                                                              |
|       |                                                              long long int
|       |                                                           %ld

Reason for change: REFPLTB-1842 test-and-diagnostic build failure in RPi4 Broadban...
Test Procedure: None.
Risks: Low.

(cherry picked from commit 5b2f1c025aca96767e66460573761aaf2a05240f)

+++ source/dmltad/cosa_wanconnectivity_operations.c
@@ -1257,7 +1257,7 @@ int send_query(struct query *query_info,struct mk_query *query_list,BOOL use_raw
 
 wait_for_response:
         start_time = clock_mono_ms();
-        WANCHK_LOG_DBG("waiting for response with timeout:%lld\n",timeout);
+        WANCHK_LOG_DBG("waiting for response with timeout:%lld\n", (long long int)timeout);
         if (timeout >= 0) {
             rc = poll(&pfd, 1, timeout);
         } else {
